requirements.txt
streamlit
langchain
sentence-transformers
pdfplumber
docx
pytesseract
pdf2image
Pillow
reportlab
requests
